backend/requirement.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
protobuf==4.22.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0
